Implemented catalog batch process lambda function
- added sqs
- subscribe lambda to sqs event
